# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /admin/users/{userId}/verification:
    put:
      operationId: updateAdminUserVerification
      tags: [admin]
      summary: Approve or revoke a user's verification badge
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
              required: [approved]
      responses:
        "200":
          description: Verification approval updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: number }
                  approved: { type: boolean }
                  approvedAt: { type: [number, "null"] }
                  approvedBy: { type: [number, "null"] }
                required: [id, approved, approvedAt, approvedBy]
        "400":
          description: Invalid user or approval value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller is not a configured system admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List the authenticated user's call history
      responses: { "200": { description: Call history, content: { application/json: { schema: { $ref: "#/components/schemas/CallList" } } } } }
    post:
      operationId: startCall
      tags: [calls]
      summary: Start an in-app audio call
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/StartCallInput" } } } }
      responses: { "201": { description: Ringing call, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } }, "400": { description: Invalid recipient, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }, "409": { description: An active call exists, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } } }
  /calls/{callId}:
    get:
      operationId: getCall
      tags: [calls]
      summary: Get a call available to the authenticated participant
      parameters: [{ $ref: "#/components/parameters/CallId" }]
      responses: { "200": { description: Call, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } }, "404": { description: Call not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } } }
  /calls/{callId}/accept:
    post:
      operationId: acceptCall
      tags: [calls]
      summary: Accept an incoming call
      parameters: [{ $ref: "#/components/parameters/CallId" }]
      responses: { "200": { description: Accepted call, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } } }
  /calls/{callId}/decline:
    post:
      operationId: declineCall
      tags: [calls]
      summary: Decline an incoming call
      parameters: [{ $ref: "#/components/parameters/CallId" }]
      responses: { "200": { description: Declined call, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } } }
  /calls/{callId}/end:
    post:
      operationId: endCall
      tags: [calls]
      summary: End an active call
      parameters: [{ $ref: "#/components/parameters/CallId" }]
      responses: { "200": { description: Ended call, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } } }
  /calls/{callId}/token:
    post:
      operationId: getCallLiveKitToken
      tags: [calls]
      summary: Get a short-lived audio token for an accepted call
      parameters: [{ $ref: "#/components/parameters/CallId" }]
      responses: { "200": { description: Participant token, content: { application/json: { schema: { $ref: "#/components/schemas/LiveKitToken" } } } }, "503": { description: LiveKit unavailable, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    CallId:
      name: callId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    StartCallInput:
      type: object
      properties: { calleeId: { type: number, minimum: 1 } }
      required: [calleeId]
    Call:
      type: object
      properties:
        id: { type: number }
        callerId: { type: number }
        calleeId: { type: number }
        status: { type: string, enum: [ringing, accepted, declined, missed, ended] }
        roomName: { type: string }
        createdAt: { type: number }
        acceptedAt: { type: [number, "null"] }
        declinedAt: { type: [number, "null"] }
        endedAt: { type: [number, "null"] }
        missedAt: { type: [number, "null"] }
        durationSeconds: { type: number }
      required: [id, callerId, calleeId, status, roomName, createdAt, acceptedAt, declinedAt, endedAt, missedAt, durationSeconds]
    CallList:
      type: object
      properties: { items: { type: array, items: { $ref: "#/components/schemas/Call" } } }
      required: [items]
    LiveKitToken:
      type: object
      properties:
        token: { type: string }
        url: { type: string }
        roomName: { type: string }
      required: [token, url, roomName]
...